flask_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0
